back-end/requirements.txt
anyio==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
cryptography==38.0.3
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.13
django-allauth==0.51.0
django-cors-headers==3.13.0
django-extensions==3.2.0
djangorestframework==3.14.0
drf-spectacular==0.24.2
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.0.0
fastjsonschema==2.16.1
h11==0.14.0
idna==3.3
importlib-metadata==5.0.0
inflection==0.5.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.8
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==0.8.4
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
oauthlib==3.2.2
outcome==1.2.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.2.1
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.6.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.0
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.10
wcwidth==0.2.5
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.3.3
wsproto==1.2.0
zipp==3.10.0
